src/third_party/blink/web_tests/platform/mac/external/wpt/pointerevents/pointerlock/pointerevent_movementxy-expected.txt

This is a testharness.js-based test.
FAIL mouse pointerevent attributes assert_equals: movementX should be the delta between current event's and last event's screenX expected 20 but got 0
Harness: the test ran to completion.
